# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BillPro - Store Billing & Reports API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: getStores
      tags: [stores]
      summary: List active and inactive stores
      responses:
        "200":
          description: Store locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createStore
      tags: [stores]
      summary: Create a store location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreInput"
      responses:
        "201":
          description: Store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /stores/{id}:
    patch:
      operationId: updateStore
      tags: [stores]
      summary: Update a store location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreUpdate"
      responses:
        "200":
          description: Store updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /stores/current:
    get:
      operationId: getCurrentStore
      tags: [stores]
      summary: Get the store selected for this employee session
      responses:
        "200":
          description: Current store context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentStore"
    patch:
      operationId: setCurrentStore
      tags: [stores]
      summary: Select the current store for this session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetCurrentStoreInput"
      responses:
        "200":
          description: Store selected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentStore"
...
  /sales:
    get:
      operationId: getSales
      tags: [sales]
      summary: List all sales
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale / invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get a sale with line items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
    patch:
      operationId: updateSale
      tags: [sales]
      summary: Update a sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleUpdate"
      responses:
        "200":
          description: Updated sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Delete a sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /sales/{id}/send-email:
    post:
      operationId: sendSaleEmail
      tags: [sales]
      summary: Send invoice by email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRequest"
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Quotations ──────────────────────────────────────────
...
components:
  schemas:
...
    LineItem:
      type: object
      required: [id, type, name, quantity, unitPrice, total]
      properties:
        id:           { type: integer }
        type:         { type: string }
        productId:    { type: ["integer", "null"] }
        serviceId:    { type: ["integer", "null"] }
        name:         { type: string }
        description:  { type: ["string", "null"] }
        quantity:     { type: number }
        unitPrice:    { type: number }
        discount:     { type: number }
        total:        { type: number }
    LineItemInput:
      type: object
      required: [type, name, quantity, unitPrice]
      properties:
        type:         { type: string }
        productId:    { type: integer }
        serviceId:    { type: integer }
        name:         { type: string }
        description:  { type: string }
        quantity:     { type: number }
        unitPrice:    { type: number }
        discount:     { type: number }

    # ── Sale ──
    Sale:
      type: object
      required: [id, invoiceNumber, status, subtotal, tax, total, createdAt]
      properties:
        id:             { type: integer }
        invoiceNumber:  { type: string }
        customerId:     { type: ["integer", "null"] }
        customerName:   { type: ["string", "null"] }
        customerEmail:  { type: ["string", "null"] }
        status:         { type: string }
        subtotal:       { type: number }
        taxRate:        { type: number }
        tax:            { type: number }
        discount:       { type: number }
        total:          { type: number }
        notes:          { type: ["string", "null"] }
        paymentMethod:  { type: ["string", "null"] }
        dueDate:        { type: ["string", "null"] }
        paidAt:         { type: ["string", "null"] }
        items:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        createdAt:      { type: string }
    SaleInput:
      type: object
      required: [items]
      properties:
        customerId:     { type: integer }
        customerName:   { type: string }
        customerEmail:  { type: string }
        status:         { type: string }
        taxRate:        { type: number }
        discount:       { type: number }
        notes:          { type: string }
        paymentMethod:  { type: string }
        dueDate:        { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/LineItemInput"
    SaleUpdate:
      type: object
      properties:
        customerId:     { type: integer }
        customerName:   { type: string }
        customerEmail:  { type: string }
        status:         { type: string }
        taxRate:        { type: number }
        discount:       { type: number }
        notes:          { type: string }
        paymentMethod:  { type: string }
        dueDate:        { type: string }
        paidAt:         { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/LineItemInput"

    # ── Quotation ──
...
    EmailRequest:
      type: object
      required: [to]
      properties:
        to:       { type: string }
        subject:  { type: string }
        message:  { type: string }
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }

    # ── Reports ──
...
    Store:
      type: object
      required: [id, name, active, isDefault, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        address: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        active: { type: boolean }
        isDefault: { type: boolean }
        createdAt: { type: string, format: date-time }
    StoreInput:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 160 }
        address: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"], format: email }
        isDefault: { type: boolean }
    StoreUpdate:
      allOf:
        - $ref: "#/components/schemas/StoreInput"
      required: []
      properties:
        active: { type: boolean }
    CurrentStore:
      type: object
      required: [storeId]
      properties:
        storeId: { type: ["integer", "null"] }
        store:
          oneOf:
            - $ref: "#/components/schemas/Store"
            - type: "null"
    SetCurrentStoreInput:
      type: object
      required: [storeId]
      additionalProperties: false
      properties:
        storeId: { type: ["integer", "null"] }
...