# lib/api-zod/src/generated/types/storeUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BillPro - Store Billing & Reports API
 * OpenAPI spec version: 0.1.0
 */
import type { StoreInput } from './storeInput';

export type StoreUpdate = StoreInput & {
  active?: boolean;
};